+++ core/src/dot/tests/format_tests.rs
@@ -31,6 +31,7 @@ fn test_build_metric() {
             thread_info_cache: None,
             total_count_restarts: 0,
             bool_stalled: false,
+            last_bool_stop: false,
             work_info: None,
         }],
         edges: vec![Edge {